+++ backend/src/main/java/com/graphy/backend/domain/auth/service/AuthService.java
@@ -3,14 +3,13 @@ package com.graphy.backend.domain.auth.service;
 import com.graphy.backend.domain.auth.controller.JwtFilter;
 import com.graphy.backend.domain.auth.dto.request.LogoutRequest;
 import com.graphy.backend.domain.auth.dto.response.GetTokenInfoResponse;
+import com.graphy.backend.domain.auth.infra.TokenProvider;
 import com.graphy.backend.domain.member.domain.Member;
 import com.graphy.backend.domain.member.dto.request.SignInMemberRequest;
 import com.graphy.backend.domain.member.dto.request.SignUpMemberRequest;
 import com.graphy.backend.domain.member.service.MemberService;
 import com.graphy.backend.domain.auth.domain.RefreshToken;
 import com.graphy.backend.domain.auth.repository.RefreshTokenRepository;
-import com.graphy.backend.global.error.ErrorCode;
-import com.graphy.backend.global.error.exception.EmptyResultException;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
