+++ backend/src/test/java/com/graphy/backend/test/MockApiTest.java
@@ -3,7 +3,7 @@ package com.graphy.backend.test;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.graphy.backend.BackendApplication;
-import com.graphy.backend.domain.auth.service.TokenProvider;
+import com.graphy.backend.domain.auth.infra.TokenProvider;
 import com.graphy.backend.domain.auth.repository.RefreshTokenRepository;
 import com.graphy.backend.test.config.TestProfile;
 import org.junit.jupiter.api.Disabled;
